Adding the dashboard with a simple ui

+++ Epub-Reader-TTS/Epub-Reader-TTS/ValueConverters/MultiplyConverter.cs
@@ -14,10 +14,9 @@ namespace Epub_Reader_TTS
     {
         public override object Convert(object value, Type targetType = null, object parameter = null, CultureInfo culture = null)
         {
-
             var multiplyer = double.Parse(parameter.ToString());
 
-            var result = int.Parse(value.ToString());
+            var result = double.Parse(value.ToString());
 
             return result * multiplyer;
